Простой Сетевой чат. 1. Добавить в сетевой чат авторизацию через базу данных SQLite. 2. *Добавить в сетевой чат возможность смены ника.
