+++ src/main/java/ru/geekbrains/server/AuthService.java
@@ -1,4 +1,4 @@
-package ru.geekbrains.sever;
+package ru.geekbrains.server;
 
 public interface AuthService {
     String getNicknameByLoginAndPassword(String login, String password);
