+++ src/main/java/ru/geekbrains/server/MainServer.java
@@ -1,4 +1,4 @@
-package ru.geekbrains.sever;
+package ru.geekbrains.server;
 
 public class MainServer {
     public static void main(String[] args) {
